requirements_for_dev.txt
flake8==7.3.0
pytest==8.3.5
pytest-django==4.10.0
pytest-cov==6.1.0
factory-boy==3.3.3
Faker==37.1.0
responses==0.25.7
freezegun==1.5.1
